% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{BetFeature}
\alias{BetFeature}
\title{Compute betweenness centrality for a list of nodes}
\usage{
BetFeature(Net, Nodes)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{Nodes}{a list of node names}
}
\value{
a list object containing the different type network features
}
\description{
This function ...
}
