% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBindingDisruption.R
\name{get_disruption_p_value}
\alias{get_disruption_p_value}
\title{Calculate transcription factor disruption p values for the SNPs}
\usage{
get_disruption_p_value(score_changes_file, bg_score_changes_dir,
  output_filename)
}
\arguments{
\item{score_changes_file}{file name containing the score changes between reference and alternate sequences around the input SNPs}

\item{bg_score_changes_dir}{directory cotaining the background score changes between the reference and alternate sequences aroung the 1000 Genome SNPs}

\item{output_filename}{the path to the file name to write the output}
}
\description{
Calculate transcription factor disruption p values for the SNPs
}
