% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModSearch.R
\name{module_search_ind}
\alias{module_search_ind}
\title{Search modules from a list of seeds/nodes}
\usage{
module_search_ind(V_weight, Adj_List, ind_seed, E_adj)
}
\arguments{
\item{V_weight}{a numeric vector of node weight}

\item{Adj_List}{a list object representing the adjacency list of the input network}

\item{E_adj}{the adjacency matrix}

\item{Net}{a graph object representing the input network}
}
\value{
a list object containing modules and related statistics
}
\description{
This function searches attached modules based on the list of seeds
}
