% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{NetFeature}
\alias{NetFeature}
\title{Compute network-based features needed for risk variant prediction}
\usage{
NetFeature(Net, Nodes)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{Nodes}{a list of node names}
}
\value{
a list object containing the different type network features
}
\description{
This function ...
}
