% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{PageFeature}
\alias{PageFeature}
\title{Compute pagerank centrality for a list of nodes}
\usage{
PageFeature(Net, edge_data)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{Nodes}{a list of node names}
}
\value{
a list object containing the different type network features
}
\description{
This function ...
}
