% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainAndPred.R
\name{predMod}
\alias{predMod}
\title{Compute prediction score}
\usage{
predMod(FeatureMatrix, RFmodel)
}
\arguments{
\item{FeatureMatrix}{a matrix of feature values}

\item{RFmodel}{a random forest classifier}
}
\value{
prediction score matrix
}
\description{
This function predicts the probablity score using a trained random forest classifier
}
