% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{snpFeature}
\alias{snpFeature}
\title{Compute TF binding disruption score}
\usage{
snpFeature(Net, eSNP_seeds)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{edge_data}{a dataframe specifying the edge information}
}
\value{
a vector of TF binding disruption scores
}
\description{
This function computes the TF binding disruption score for a list of candidate snps
}
