% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainAndPred.R
\name{trainMod}
\alias{trainMod}
\title{Train a classifier}
\usage{
trainMod(FeatureMatrix)
}
\arguments{
\item{FeatureMatrix}{a matrix of feature values}

\item{pos}{a vector specifying the positive set of elements}

\item{neg}{a vector specifying the control set of elements}
}
\value{
a random forest classifier
}
\description{
This function ...
}
