% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{WDFeature}
\alias{WDFeature}
\title{Compute weighted degree for a list of nodes}
\usage{
WDFeature(Adj_List, edge_data)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{Nodes}{a list of node names}
}
\value{
a list object containing the different type network features
}
\description{
This function ...
}
