% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{BetFeature}
\alias{BetFeature}
\title{Compute betweenness centrality}
\usage{
BetFeature(Net, edge_data)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{edge_data}{a dataframe specifying the edge information}
}
\value{
a vector of betweenness centrality values
}
\description{
This function computes the betweenness centrality for a list of candidate snps
}
