% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{CloseFeature}
\alias{CloseFeature}
\title{Compute closeness centrality}
\usage{
CloseFeature(Net, edge_data)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{edge_data}{a dataframe specifying the edge information}
}
\value{
a vector of closeness centrality values
}
\description{
This function computes the closeness centrality for a list of candidate snps
}
