% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBindingDisruption.R
\name{getTFname}
\alias{getTFname}
\title{Extract TF name from the file name which is separated from motif name with @: AHR@M6139_1.02}
\usage{
getTFname(filename)
}
\arguments{
\item{filename}{file name containing the score changes between reference and alternate sequences around the input SNPs}
}
\description{
Extract TF name from the file name which is separated from motif name with @: AHR@M6139_1.02
}
