% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetConfig.R
\name{makeNet}
\alias{makeNet}
\title{Generate a graph object for input network}
\usage{
makeNet(edgeFile, nodeFile, rand = FALSE, seed = 1)
}
\arguments{
\item{rand}{boolean value indicating if randomized network will be generated}

\item{seed}{seed for network randomization}

\item{Interaction_file}{file path for edges in input network}

\item{Node_file}{file path for nodes in input network}
}
\value{
a graph object
}
\description{
This function ...
}
