% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetConfig.R
\name{makeNet}
\alias{makeNet}
\title{Generate a graph object for input network}
\usage{
makeNet(edgeFile, nodeFile, rand = FALSE, seed = 1)
}
\arguments{
\item{rand}{boolean value indicating if randomized network will be generated}

\item{seed}{seed for network randomization}

\item{edge_file}{file path for edges in input network}

\item{node_file}{file path for nodes in input network}
}
\value{
a graph/network object(igraph package)
}
\description{
This function reads a node file and edge file and generates a network object for downstream analysis.
}
