% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{ModuleFeature}
\alias{ModuleFeature}
\title{Compute module scores}
\usage{
ModuleFeature(Adj_List, E_adj, eSNP_seeds, V_weight, Nodes)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{edge_data}{a dataframe specifying the edge information}
}
\value{
a vector of module scores
}
\description{
This function computes the module scores for a list of candidate snps
}
