% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{NetFeature}
\alias{NetFeature}
\title{Compute network-based features needed for risk variant prediction}
\usage{
NetFeature(Net, nodeFile, edgeFile, snpFile)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{edgeFile}{the path for the network file}

\item{Nodes}{the path for the node file}
}
\value{
a matrix containing the different type network features
}
\description{
This function ...
}
