% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{PageFeature}
\alias{PageFeature}
\title{Compute pagerank centrality}
\usage{
PageFeature(Net, edge_data)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{edge_data}{a dataframe specifying the edge information}
}
\value{
a vector of pagerank centrality values
}
\description{
This function computes the pagerank centrality for a list of candidate snps
}
