% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{snpFeature}
\alias{snpFeature}
\title{Compute module score for a list of nodes}
\usage{
snpFeature(Net, eSNP_seeds)
}
\arguments{
\item{Net}{a graph object representing the input network}

\item{Nodes}{a list of node names}

\item{Net}{a graph object representing the input network}

\item{Nodes}{a list of node names}
}
\value{
a list object containing the different type network features 

Compute pagerank centrality for a list of nodes

This function ...

a list object containing the different type network features
}
\description{
This function ...
}
