% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainAndPred.R
\name{trainMod}
\alias{trainMod}
\title{Train a classifier}
\usage{
trainMod(FeatureMatrix)
}
\arguments{
\item{FeatureMatrix}{a matrix of feature values plus the last column indicating positive or negative snps}
}
\value{
a random forest classifier
}
\description{
This function trains a random forest classifier using different types of features
}
