% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFeatureCalculation.R
\name{WDFeature}
\alias{WDFeature}
\title{Compute weighted degree}
\usage{
WDFeature(Adj_List, edge_data)
}
\arguments{
\item{edge_data}{a dataframe specifying the edge information}

\item{Net}{a graph object representing the input network}
}
\value{
a vector of weighted degree values
}
\description{
This function computes the weighted degree for a list of candidate snps
}
